SLING-12858 refactor to use the felix.http.wrappers
