--- src/main/java/org/apache/sling/servlethelpers/BaseResponseBodySupport.java
@@ -1,87 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.sling.servlethelpers;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.StandardCharsets;
-import java.util.Objects;
-
-/**
- * Manage response body content.
- */
-public abstract class BaseResponseBodySupport<T extends OutputStream> {
-
-    protected ByteArrayOutputStream outputStream;
-    protected T servletOutputStream;
-    protected PrintWriter printWriter;
-
-    protected BaseResponseBodySupport() {
-        reset();
-    }
-
-    public void reset() {
-        outputStream = new ByteArrayOutputStream();
-        servletOutputStream = null;
-        printWriter = null;
-    }
-
-    public abstract T getOutputStream();
-
-    public PrintWriter getWriter(String charset) {
-        if (printWriter == null) {
-            try {
-                printWriter = new PrintWriter(new OutputStreamWriter(getOutputStream(), defaultCharset(charset)));
-            } catch (UnsupportedEncodingException ex) {
-                throw new RuntimeException("Unsupported encoding: " + defaultCharset(charset), ex);
-            }
-        }
-        return printWriter;
-    }
-
-    public byte[] getOutput() {
-        if (printWriter != null) {
-            printWriter.flush();
-        }
-        if (servletOutputStream != null) {
-            try {
-                servletOutputStream.flush();
-            } catch (IOException ex) {
-                // ignore
-            }
-        }
-        return outputStream.toByteArray();
-    }
-
-    public String getOutputAsString(String charset) {
-        try {
-            return new String(getOutput(), defaultCharset(charset));
-        } catch (UnsupportedEncodingException ex) {
-            throw new RuntimeException("Unsupported encoding: " + defaultCharset(charset), ex);
-        }
-    }
-
-    private String defaultCharset(String charset) {
-        return Objects.toString(charset, StandardCharsets.UTF_8.name());
-    }
-}
