+++ src/main/java/org/apache/sling/servlethelpers/JakartaByteArrayPart.java
@@ -18,12 +18,68 @@
  */
 package org.apache.sling.servlethelpers;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Collection;
+
 import jakarta.servlet.http.Part;
 
 /**
  * Simple Part implementation backed by an in-memory byte array
  */
-public class JakartaByteArrayPart extends BaseByteArrayPart implements Part {
+public class JakartaByteArrayPart implements Part {
+    private final byte[] content;
+    private final String name;
+
+    JakartaByteArrayPart(byte[] content, String name) {
+        if (content == null) throw new IllegalArgumentException("content may not be null");
+
+        if (name == null || name.isEmpty()) throw new IllegalArgumentException("name may not be null or empty");
+
+        this.content = content;
+        this.name = name;
+    }
+
+    public InputStream getInputStream() throws IOException { // NOSONAR
+        return new ByteArrayInputStream(content);
+    }
+
+    public String getContentType() {
+        return null;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getSubmittedFileName() {
+        return getName();
+    }
+
+    public long getSize() {
+        return content.length;
+    }
+
+    public void write(String fileName) throws IOException {
+        throw new UnsupportedOperationException();
+    }
+
+    public void delete() throws IOException {
+        throw new UnsupportedOperationException();
+    }
+
+    public String getHeader(String name) {
+        throw new UnsupportedOperationException();
+    }
+
+    public Collection<String> getHeaders(String name) {
+        throw new UnsupportedOperationException();
+    }
+
+    public Collection<String> getHeaderNames() {
+        throw new UnsupportedOperationException();
+    }
 
     /**
      * Returns a Builder instance used to create a ByteArrayPart
@@ -34,13 +90,20 @@ public class JakartaByteArrayPart extends BaseByteArrayPart implements Part {
         return new Builder();
     }
 
-    private JakartaByteArrayPart(byte[] content, String name) {
-        super(content, name);
-    }
+    public static class Builder {
+        protected byte[] content;
+        protected String name;
 
-    public static class Builder extends BaseBuilder<JakartaByteArrayPart> {
+        public Builder withContent(byte[] content) {
+            this.content = content;
+            return this;
+        }
 
-        @Override
+        public Builder withName(String name) {
+            this.name = name;
+            return this;
+        }
+
         public JakartaByteArrayPart build() {
             return new JakartaByteArrayPart(content, name);
         }
