+++ src/main/java/org/apache/sling/servlethelpers/MockServletContext.java
@@ -18,23 +18,13 @@
  */
 package org.apache.sling.servlethelpers;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterRegistration;
-import javax.servlet.RequestDispatcher;
 import javax.servlet.Servlet;
 import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-import javax.servlet.ServletRegistration.Dynamic;
-import javax.servlet.SessionCookieConfig;
-import javax.servlet.SessionTrackingMode;
 import javax.servlet.descriptor.JspConfigDescriptor;
 
 import java.util.Enumeration;
-import java.util.EventListener;
-import java.util.Map;
-import java.util.Set;
 
+import org.apache.felix.http.javaxwrappers.ServletContextWrapper;
 import org.osgi.annotation.versioning.ConsumerType;
 
 /**
@@ -44,21 +34,10 @@ import org.osgi.annotation.versioning.ConsumerType;
  */
 @ConsumerType
 @Deprecated(since = "2.0.0")
-public class MockServletContext extends BaseMockServletContext implements ServletContext {
+public class MockServletContext extends ServletContextWrapper {
 
-    @Override
-    public ServletContext getContext(final String uriPath) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public RequestDispatcher getNamedDispatcher(final String name) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public RequestDispatcher getRequestDispatcher(final String path) {
-        throw new UnsupportedOperationException();
+    public MockServletContext(MockJakartaServletContext wrappedServletContext) {
+        super(wrappedServletContext);
     }
 
     @Override
@@ -73,106 +52,6 @@ public class MockServletContext extends BaseMockServletContext implements Servle
 
     @Override
     public Enumeration<Servlet> getServlets() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void log(final Exception exception, final String msg) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Dynamic addServlet(final String servletName, final String className) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Dynamic addServlet(final String servletName, final Servlet servlet) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Dynamic addServlet(final String servletName, final Class<? extends Servlet> servletClass) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T extends Servlet> T createServlet(final Class<T> clazz) throws ServletException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public ServletRegistration getServletRegistration(final String servletName) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Map<String, ? extends ServletRegistration> getServletRegistrations() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public FilterRegistration.Dynamic addFilter(final String filterName, final String className) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public FilterRegistration.Dynamic addFilter(final String filterName, final Filter filter) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public FilterRegistration.Dynamic addFilter(final String filterName, final Class<? extends Filter> filterClass) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T extends Filter> T createFilter(final Class<T> clazz) throws ServletException {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public FilterRegistration getFilterRegistration(final String filterName) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Map<String, ? extends FilterRegistration> getFilterRegistrations() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public SessionCookieConfig getSessionCookieConfig() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void setSessionTrackingModes(final Set<SessionTrackingMode> sessionTrackingModes) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<SessionTrackingMode> getDefaultSessionTrackingModes() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<SessionTrackingMode> getEffectiveSessionTrackingModes() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T extends EventListener> void addListener(final T listener) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void addListener(final Class<? extends EventListener> listenerClass) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T extends EventListener> T createListener(final Class<T> clazz) {
         throw new UnsupportedOperationException();
     }
