--- src/main/java/org/apache/sling/servlethelpers/ResponseBodySupport.java
@@ -1,60 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.sling.servlethelpers;
-
-import javax.servlet.ServletOutputStream;
-import javax.servlet.WriteListener;
-
-import java.io.IOException;
-
-/**
- * Manage response body content.
- *
- * @deprecated Use {@link JakartaResponseBodySupport} instead.
- */
-@Deprecated(since = "2.0.0")
-class ResponseBodySupport extends BaseResponseBodySupport<ServletOutputStream> {
-
-    public ResponseBodySupport() {
-        reset();
-    }
-
-    @Override
-    public ServletOutputStream getOutputStream() {
-        if (servletOutputStream == null) {
-            servletOutputStream = new ServletOutputStream() {
-                @Override
-                public void write(int b) throws IOException {
-                    outputStream.write(b);
-                }
-
-                @Override
-                public boolean isReady() {
-                    return true;
-                }
-
-                @Override
-                public void setWriteListener(WriteListener writeListener) {
-                    throw new UnsupportedOperationException();
-                }
-            };
-        }
-        return servletOutputStream;
-    }
-}
